Update Entry and Editor to use Windows 11 styles
