+++ src/Compatibility/Core/src/Windows/Resources.xaml
@@ -350,10 +350,6 @@
 		</Setter>
 	</Style>
 
-	<Style TargetType="TextBox">
-		<Setter Property="Margin" Value="0" />
-	</Style>
-
 	<Style TargetType="ToggleSwitch">
 		<Setter Property="MinWidth" Value="0"/>
 	</Style>
