+++ src/Core/src/Handlers/Entry/EntryHandler.Android.cs
@@ -63,7 +63,6 @@ namespace Microsoft.Maui.Handlers
 		void OnTextChanged(object? sender, Android.Text.TextChangedEventArgs e) =>
 			VirtualView.UpdateText(e);
 
-		// This is a Android-specific mapping
 		public static void MapBackground(EntryHandler handler, IEntry entry)
 		{
 			handler.NativeView?.UpdateBackground(entry);
