+++ src/Core/src/Platform/Windows/TextBoxExtensions.cs
@@ -1,107 +1,122 @@
-using System.Collections.Generic;
+#nullable enable
 using Microsoft.Maui.Graphics;
-using Microsoft.UI.Xaml.Media;
+using Microsoft.Maui.Platform;
+using Microsoft.UI.Xaml.Controls;
 
 namespace Microsoft.Maui
 {
 	public static class TextBoxExtensions
 	{
-		public static void UpdateText(this MauiTextBox nativeControl, IEditor editor)
+		public static void UpdateText(this TextBox nativeControl, ITextInput textInput)
 		{
-			string newText = editor.Text;
+			var newText = textInput.Text;
 
 			if (nativeControl.Text == newText)
 				return;
 
-			nativeControl.Text = newText;
+			nativeControl.Text = newText ?? string.Empty;
 
 			if (!string.IsNullOrEmpty(nativeControl.Text))
 				nativeControl.SelectionStart = nativeControl.Text.Length;
 		}
 
-		public static void UpdateText(this MauiTextBox textBox, IEntry entry)
+		public static void UpdateBackground(this TextBox textBox, IView view)
 		{
-			textBox.Text = entry.Text ?? string.Empty;
+			var brush = view.Background?.ToNative();
+			if (brush is null)
+			{
+				textBox.Resources.Remove("TextControlBackground");
+				textBox.Resources.Remove("TextControlBackgroundPointerOver");
+				textBox.Resources.Remove("TextControlBackgroundFocused");
+				textBox.Resources.Remove("TextControlBackgroundDisabled");
+			}
+			else
+			{
+				textBox.Resources["TextControlBackground"] = brush;
+				textBox.Resources["TextControlBackgroundPointerOver"] = brush;
+				textBox.Resources["TextControlBackgroundFocused"] = brush;
+				textBox.Resources["TextControlBackgroundDisabled"] = brush;
+			}
 		}
 
-		public static void UpdateTextColor(this MauiTextBox textBox, ITextStyle textStyle)
+		public static void UpdateTextColor(this TextBox textBox, ITextStyle textStyle)
 		{
-			if (textStyle.TextColor == null)
-				return;
-
-			var brush = textStyle.TextColor.ToNative();
-
-			textBox.Foreground = brush;
-			textBox.ForegroundFocusBrush = brush;
+			var brush = textStyle.TextColor?.ToNative();
+			if (brush is null)
+			{
+				textBox.Resources.Remove("TextControlForeground");
+				textBox.Resources.Remove("TextControlForegroundPointerOver");
+				textBox.Resources.Remove("TextControlForegroundFocused");
+				textBox.Resources.Remove("TextControlForegroundDisabled");
+			}
+			else
+			{
+				textBox.Resources["TextControlForeground"] = brush;
+				textBox.Resources["TextControlForegroundPointerOver"] = brush;
+				textBox.Resources["TextControlForegroundFocused"] = brush;
+				textBox.Resources["TextControlForegroundDisabled"] = brush;
+			}
 		}
 
-		public static void UpdateCharacterSpacing(this MauiTextBox textBox, ITextStyle textStyle)
+		public static void UpdateCharacterSpacing(this TextBox textBox, ITextStyle textStyle)
 		{
 			textBox.CharacterSpacing = textStyle.CharacterSpacing.ToEm();
 		}
 
-		public static void UpdateCharacterSpacing(this MauiTextBox textBox, IEntry entry)
+		public static void UpdateReturnType(this TextBox textBox, ITextInput textInput)
 		{
-			textBox.CharacterSpacing = entry.CharacterSpacing.ToEm();
+			textBox.UpdateInputScope(textInput);
 		}
 
-		public static void UpdateReturnType(this MauiTextBox textBox, IEntry entry)
+		public static void UpdateClearButtonVisibility(this TextBox textBox, IEntry entry)
 		{
-			textBox.UpdateInputScope(entry);
+			if (textBox is IMauiTextBox mauiEntryTextBox)
+				mauiEntryTextBox.IsDeleteButtonEnabled = entry.ClearButtonVisibility == ClearButtonVisibility.WhileEditing;
 		}
 
-		public static void UpdateClearButtonVisibility(this MauiTextBox textBox, IEntry entry)
+		public static void UpdatePlaceholder(this TextBox textBox, IPlaceholder placeholder)
 		{
-			textBox.ClearButtonVisible = entry.ClearButtonVisibility == ClearButtonVisibility.WhileEditing;
+			textBox.PlaceholderText = placeholder.Placeholder ?? string.Empty;
 		}
 
-		public static void UpdatePlaceholder(this MauiTextBox textBox, IEditor editor)
+		public static void UpdatePlaceholderColor(this TextBox textBox, IPlaceholder placeholder)
 		{
-			textBox.PlaceholderText = editor.Placeholder ?? string.Empty;
+			var brush = placeholder.PlaceholderColor?.ToNative();
+			if (brush is null)
+			{
+				// Windows.Foundation.UniversalApiContract < 5
+				textBox.Resources.Remove("TextControlPlaceholderForeground");
+				textBox.Resources.Remove("TextControlPlaceholderForegroundPointerOver");
+				textBox.Resources.Remove("TextControlPlaceholderForegroundFocused");
+				textBox.Resources.Remove("TextControlPlaceholderForegroundDisabled");
+
+				// Windows.Foundation.UniversalApiContract >= 5
+				textBox.ClearValue(TextBox.PlaceholderForegroundProperty);
+			}
+			else
+			{
+				// Windows.Foundation.UniversalApiContract < 5
+				textBox.Resources["TextControlPlaceholderForeground"] = brush;
+				textBox.Resources["TextControlPlaceholderForegroundPointerOver"] = brush;
+				textBox.Resources["TextControlPlaceholderForegroundFocused"] = brush;
+				textBox.Resources["TextControlPlaceholderForegroundDisabled"] = brush;
+
+				// Windows.Foundation.UniversalApiContract >= 5
+				textBox.PlaceholderForeground = brush;
+			}
 		}
 
-		public static void UpdatePlaceholder(this MauiTextBox textBox, IEntry entry)
-		{
-			textBox.PlaceholderText = entry.Placeholder ?? string.Empty;
-		}
-	
-		public static void UpdatePlaceholderColor(this MauiTextBox textBox, IPlaceholder placeholder, Brush? defaultPlaceholderColorBrush, Brush? defaultPlaceholderColorFocusBrush)
-		{
-			Color placeholderColor = placeholder.PlaceholderColor;
-
-			BrushHelpers.UpdateColor(placeholderColor, ref defaultPlaceholderColorBrush,
-				() => textBox.PlaceholderForegroundBrush, brush => textBox.PlaceholderForegroundBrush = brush);
-
-			BrushHelpers.UpdateColor(placeholderColor, ref defaultPlaceholderColorFocusBrush,
-				() => textBox.PlaceholderForegroundFocusBrush, brush => textBox.PlaceholderForegroundFocusBrush = brush);
-		}
-
-		public static void UpdateFont(this MauiTextBox nativeControl, IText text, IFontManager fontManager) =>
+		public static void UpdateFont(this TextBox nativeControl, IText text, IFontManager fontManager) =>
 			nativeControl.UpdateFont(text.Font, fontManager);
 
-		public static void UpdateIsReadOnly(this MauiTextBox textBox, IEditor editor)
+		public static void UpdateIsReadOnly(this TextBox textBox, ITextInput textInput)
 		{
-			textBox.IsReadOnly = editor.IsReadOnly;
+			textBox.IsReadOnly = textInput.IsReadOnly;
 		}
 
-		public static void UpdateIsReadOnly(this MauiTextBox textBox, IEntry entry)
+		public static void UpdateMaxLength(this TextBox textBox, ITextInput textInput)
 		{
-			textBox.IsReadOnly = entry.IsReadOnly;
-		}
-
-		public static void UpdateMaxLength(this MauiTextBox textBox, IEditor editor)
-		{
-			textBox.MaxLength = editor.MaxLength;
-
-			var currentControlText = textBox.Text;
-
-			if (currentControlText.Length > editor.MaxLength)
-				textBox.Text = currentControlText.Substring(0, editor.MaxLength);
-		}
-
-		public static void UpdateMaxLength(this MauiTextBox textBox, IEntry entry)
-		{
-			var maxLength = entry.MaxLength;
+			var maxLength = textInput.MaxLength;
 
 			if (maxLength == -1)
 				maxLength = int.MaxValue;
@@ -114,22 +129,17 @@ namespace Microsoft.Maui
 				textBox.Text = currentControlText.Substring(0, maxLength);
 		}
 
-		public static void UpdateIsPassword(this MauiTextBox textBox, IEntry entry)
-		{
-			textBox.IsPassword = entry.IsPassword;
-		}
-
-		public static void UpdateIsTextPredictionEnabled(this MauiTextBox textBox, ITextInput textInput)
+		public static void UpdateIsTextPredictionEnabled(this TextBox textBox, ITextInput textInput)
 		{
 			textBox.UpdateInputScope(textInput);
 		}
 
-		public static void UpdateKeyboard(this MauiTextBox textBox, ITextInput textInput)
+		public static void UpdateKeyboard(this TextBox textBox, ITextInput textInput)
 		{
 			textBox.UpdateInputScope(textInput);
 		}
 
-		internal static void UpdateInputScope(this MauiTextBox textBox, ITextInput textInput)
+		internal static void UpdateInputScope(this TextBox textBox, ITextInput textInput)
 		{
 			if (textInput.Keyboard is CustomKeyboard custom)
 			{
@@ -152,7 +162,7 @@ namespace Microsoft.Maui
 			textBox.InputScope = inputScope;
 		}
 
-		public static void UpdateHorizontalTextAlignment(this MauiTextBox textBox, ITextAlignment textAlignment)
+		public static void UpdateHorizontalTextAlignment(this TextBox textBox, ITextAlignment textAlignment)
 		{
 			// We don't have a FlowDirection yet, so there's nothing to pass in here. 
 			// TODO: Update this when FlowDirection is available 
@@ -160,9 +170,22 @@ namespace Microsoft.Maui
 			textBox.TextAlignment = textAlignment.HorizontalTextAlignment.ToNative(true);
 		}
 
-		public static void UpdateVerticalTextAlignment(this MauiTextBox textBox, ITextAlignment textAlignment)
+		public static void UpdateVerticalTextAlignment(this TextBox textBox, ITextAlignment textAlignment)
 		{
-			textBox.VerticalAlignment = textAlignment.VerticalTextAlignment.ToNativeVerticalAlignment();
+			if (textBox is IMauiTextBox mauiEntryTextBox)
+				mauiEntryTextBox.VerticalTextAlignment = textAlignment.VerticalTextAlignment.ToNativeVerticalAlignment();
+		}
+
+		public static void UpdateCursorPosition(this TextBox textBox, IEntry entry)
+		{
+			if (textBox.SelectionStart != entry.CursorPosition)
+				textBox.SelectionStart = entry.CursorPosition;
+		}
+
+		public static void UpdateSelectionLength(this TextBox textBox, IEntry entry)
+		{
+			if (textBox.SelectionLength != entry.SelectionLength)
+				textBox.SelectionLength = entry.SelectionLength;
 		}
 	}
 }
